Move upload_file into Server
